Split guardian_secret|public_key.rs
